split Python package into 2 and 3
This allows linking both into a view side by side. It breaks a few
things too:

- packages which depended on either version of Python are now forced to
  depend on 2
- packages extending either version of Python now extend only 2

Neither should be that big of a problem though, especially since both
versions were changed so as to install pip and `pip install --user` (or
virtualenvs) is the recommended way (for us at least) to install new
libraries.